pyproject.toml
[project]
name = "concert-playlist-llm"
version = "0.1.0"
description = "LLM backend module for intent parsing and track tagging"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.0",
    "fastapi>=0.112.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.4.0",
    "pyjwt>=2.9.0",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
